requirements.txt
streamlit==1.45.1
transformers==4.52.3
fastapi==0.115.12
uvicorn==0.34.2
peft==0.15.2
python-dotenv==1.1.0